% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggRaster.R
\name{aggRaster}
\alias{aggRaster}
\title{Temporal aggregation of raster data}
\usage{
aggRaster(data, agg.level, write.file = TRUE, days, path.out,
  process.incomplete = FALSE, ...)
}
\arguments{
\item{data}{List of raster files with each list entry containing
all hourly raster files of one single day.}

\item{agg.level}{Aggregation level in hours (numeric).}

\item{write.file}{logical: Should aggregated raster files be locally stored
(TRUE) or kept in memory (FALSE).}

\item{days}{character: Days with available hourly raster files.}

\item{path.out}{character: Path to output directory.}

\item{process.incomplete}{logical: Should incomplete raster data sets (e.g. two raster 
files at agg.level = 3) be processed?}

\item{...}{Further arguments to be passed.}
}
\value{
A raster Layer
}
\description{
Temporal aggregation of raster data
}
\author{
Meike Kuehnlein
}

