% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/borgIndices.R
\name{borgIndices}
\alias{borgIndices}
\title{Calculate shape indices listed in Borg 1998}
\usage{
borgIndices(Ar, Ur, De, Du)
}
\arguments{
\item{Ar}{area of the entities}

\item{Ur}{Perimeter of the cloud entities}

\item{De}{Diameter of the inner circle}

\item{Du}{Diameter of the outer circle}
}
\value{
A RasterStack
}
\description{
Calculate shape indices listed in Borg 1998
}
\details{
This function is mainly used as internal function of
\code{\link{geometryVariables}}
}
\examples{
msg_example <-  raster(system.file("msg",
 "201010081250_mt09s_ct01dk009_m1hct_1000_rg01de_003000.rst",
 package="Rainfall"))
 # set non clouded areas to NA:
 msg_example=reclassify(msg_example, cbind(-99,NA))
 #calculate geometry Variables
 geometry <- geometryVariables(msg_example,var=c("Ar","Ur","Re","Ru"))
 borg<-borgIndices(geometry$Ar,geometry$Ur,geometry$Re*2, geometry$Ru*2)
 plot(borg)
}
\author{
Hanna Meyer
}
\references{
Borg, E. & Fichtelmann, B. (1998):
Vergleichende Analyse von Formindizes zur Charaktersierung von
Landschaftsobjekten unter ökologischen Aspekten.
 Zeitschrift für Photogrammetrie und Vermessung, 66, 108-112.
}

