% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcScalingStats.R
\name{calcScalingStats}
\alias{calcScalingStats}
\title{Returns mean and sd values of the predictors which can be used for scaling
predictors used for training}
\usage{
calcScalingStats(x)
}
\arguments{
\item{x}{Data.frame or rasterStack of the predictors}
}
\value{
Data.frame with mean and sd values
}
\description{
Returns mean and sd values of the predictors which can be used for scaling
predictors used for training
}
\examples{
msg_example <-getChannels(inpath=system.file("extdata/msg",package="Rainfall"))
scaleparam <- calcScalingStats(msg_example)
}
\author{
Hanna Meyer
}

