% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculatePredictors.R
\name{calculatePredictors}
\alias{calculatePredictors}
\title{Calculate predictors from MSG cloud masked data}
\usage{
calculatePredictors(scenerasters, sunzenith, variables, derivVariables,
  xderivTexture, date)
}
\arguments{
\item{scenerasters}{A raster stack of cloud masked MSG scenes with non
clouded areas were set to NA}

\item{sunzenith}{A raster of the sun zenith values}

\item{variables}{A character vector indicating the msg channels to be included.
Possible values: "VIS0.6","VIS0.8","NIR1.6","IR3.9","WV6.2","WV7.3","IR8.7",
"IR9.7","IR10.8","IR12.0","IR13.4".}

\item{derivVariables}{A character vector of derivated variables
 to be calculated. possible values are
"T0.6_1.6","T6.2_10.8","T7.3_12.0","T8.7_10.8","T10.8_12.0",
"T3.9_7.3","T3.9_10.8","sunzenith")}

\item{xderivTexture}{MSG channels from which texture is to be calculated}

\item{date}{Date of the msg scene in format yyyymmddhhmm}
}
\description{
Calculate predictors from MSG cloud masked data
}
\examples{
#list msg raster in the example folder:
scenes<-list.files(system.file("msg",package="Rainfall"),
pattern=".rst$")

# name the variables (according file name convention)
#which are to be used:
x=c("ca02p0001","ca02p0002","ca02p0003","ct01dk004","ct01dk005",
"ct01dk006","ct01dk007","ct01dk008","ct01dk009","ct01dk010","ct01dk011")

# raster the sunzenith raster which is defined as "ma11" according to
#file name conventions:
sunzenith<-raster(system.file("msg",
scenes[substr(scenes,20,23) =="ma11"],package="Rainfall"))

# stack the msg scenes:
msg_example <-  stack(system.file("msg",
scenes[substr(scenes,20,28)\%in\%x],package="Rainfall"))

date <- substr(scenes[1],1,12)

# set non clouded areas to NA:
msg_example=reclassify(msg_example, cbind(-99,NA))

# name the msg channels:
names(msg_example)<-c("VIS0.6","VIS0.8","NIR1.6","IR3.9",
"WV6.2","WV7.3","IR8.7","IR9.7","IR10.8","IR12.0","IR13.4")

#calculate variables (takes some time...)
pred <- calculatePredictors(msg_example,sunzenith,variables=c("VIS0.6","VIS0.8","NIR1.6"),
derivVariables=c("T0.6_1.6","T6.2_10.8"),
xderivTexture=c("VIS0.6","T0.6_1.6"),date=date)
print(pred)
}

