% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cr2ll.R
\name{cr2ll}
\alias{cr2ll}
\title{Get geocoordinates from MSG column and rows}
\usage{
cr2ll(colrow, ccoff = 1856, lloff = 1856)
}
\arguments{
\item{ccoff}{Integer coefficient of the scalling function (see page 28, Ref [1])}

\item{lloff}{Integer coefficient of the scalling function  (see page 28, Ref [1])}

\item{column}{Integer}

\item{row}{Integer}
}
\value{
Data frame including latitude and longitude of the MSG image
}
\description{
Get geocoordinates from MSG column and rows
}
\examples{
ccr2ll(data.frame(c(2899,3435),c(1200,1340)))
}
\author{
Hanna Meyer based on the FORTRAN Routine of EUMETSAT 2005, 2009.
}
\references{
[1] LRIT/HRIT Global Specification
(CGMS 03, Issue 2.6, 12.08.1999)
 for the parameters used in the program.

 [2] MSG Ground Segment LRIT/HRIT Mission Specific
 Implementation, EUMETSAT Document,
 (EUM/MSG/SPE/057, Issue 6, 21. June 2006).
}

