% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{download_rain}
\alias{download_rain}
\title{Download monthly rainfall data for countries}
\usage{
download_rain(country = NULL, long = T, delete_raw = T)
}
\arguments{
\item{country}{Character vector of ISO3 codes for the countries for which
rainfall data is to be downloaded. I suggest using the excellent
\code{\href{https://cran.r-project.org/web/packages/countrycode/index.html}{countrycode}}
package to retrieve ISO3 codes from country names.}

\item{long}{Logical. If TRUE, the resulting data frame is a tidy output in which
every observation is a country-year month of rainfall data. If FALSE every row is a
year with the JAN to DEC monthly values as columns.}

\item{delete_raw}{Logical. Delete local copies of the downloaded CRU files?
Defaults to TRUE.}
}
\value{
If long == TRUE \code{download_rain()} returns a data frame / tibble with the following
variables:
\itemize{
\item \code{iso3c}: ISO3 code of a country
\item \code{YEAR}: Year (currently 1901 - 2015)
\item \code{month}: Month (character string of month, JAN - DEC)
\item \code{month_numeric}: Month (numeric, 1-12)
\item \code{rain}: Monthly mean of rainfall in milimeters per month (https://crudata.uea.ac.uk/cru/data/hrg/#info)
\item \code{MAM, JJA, SON, DJF}: seasonal sum
\item \code{ANN}: Annual sum of mean rainfall per month
}
}
\description{
\code{download_rain()} downloads monthly rainfall data from the
Climate Research Unit at the University of East Anglia (\url{http://www.cru.uea.ac.uk/}).
}
\examples{

\dontrun{
# Download rainfall data for the Democratic Republic of the Congo.
rain_drc <- download_rain(country = "COD", long = T, delete_raw = T)

}

}
\references{
Harris, I, PD Jones, TJ Osborn & DH Lister (2014)
Updated high-resolution grids of monthly climatic observations - the
CRU TS3.10 Dataset. \emph{International Journal of Climatology} 34(3): 623-642.
}

