% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fourStats.R
\name{fourStats}
\alias{fourStats}
\title{nnet with cutoff as additional tuning parameter}
\usage{
fourStats(data, lev = levels(data$obs), model = NULL)
}
\description{
is used as internal function in caret's rfe and train
}
\author{
Max Kuhn
}
\references{
http://caret.r-forge.r-project.org/custom_models.html
}

