% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateFileName.R
\name{generateFileName}
\alias{generateFileName}
\title{Generate msg file names}
\usage{
generateFileName(x, date)
}
\arguments{
\item{x}{Raster stack containing the MSG scenes}

\item{date}{see getDate}
}
\value{
a vector of character strings
}
\description{
Generate msg file names
}
\examples{
msg_example <-getChannels(inpath=system.file("extdata/msg",package="Rainfall"),
channels=c("VIS0.6","IR12.0"))
generateFileName(msg_example,getDate(inpath=system.file("extdata/msg",package="Rainfall")))
}
\author{
Hanna Meyer
}

