% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geometryVariables.R
\name{geometryVariables}
\alias{geometryVariables}
\title{Calculate selected geometry variables for clouds}
\usage{
geometryVariables(sceneraster, var = c("Ar"), setNA = -99)
}
\arguments{
\item{sceneraster}{A rasterLayer containing NA for non clouds and any value for
clouded areas}
}
\value{
A list of RasterStacks containing the texture parameters for each
combination of channel and filter
}
\description{
Calculate selected geometry variables for clouds
}
\examples{
msg_example <-getChannels(inpath=system.file("msg",package="Rainfall"),
channels="VIS0.8")

#calculate geometry Variables
geometry <- geometryVariables(msg_example,var=c("Ar","Ur"))
plot(geometry)
}
\author{
Hanna Meyer
}
\seealso{
\code{\link{SDMTools}}, \code{\link{clump}},
 \code{\link{borgIndices}}
}

