% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChannels.R
\name{getChannels}
\alias{getChannels}
\title{Get channels from a msg scene's inputpath}
\usage{
getChannels(inpath, type = "rst", channels = c("VIS0.6", "VIS0.8", "NIR1.6",
  "IR3.9", "WV6.2", "WV7.3", "IR8.7", "IR9.7", "IR10.8", "IR12.0", "IR13.4"),
  mask = -99)
}
\arguments{
\item{inpath}{The path to the msg channels of one scene}

\item{type}{A character string indicating the file type}

\item{channels}{A character vector of the channels to return}

\item{mask}{A numeric number which is to bet set to NA 
because there are no clouds.}
}
\value{
A Raster stack of the MSG channels
}
\description{
Get channels from a msg scene's inputpath
}
\examples{
msg_example <-getChannels(inpath=system.file("extdata/msg",package="Rainfall"),
channels=c("VIS0.6","IR12.0"))
}
\author{
Hanna Meyer
}

