% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getDate.R
\name{getDate}
\alias{getDate}
\title{Get the date from a msg scene's inputpath}
\usage{
getDate(inpath, type = "rst")
}
\arguments{
\item{inpath}{The path to the msg channels including sunzenith of one scene}

\item{type}{A character string indicating the file type}
}
\value{
The date as character
}
\description{
Get the date from a msg scene's inputpath
}
\examples{
date <-getDate(inpath=system.file("msg",package="Rainfall"))
}
\author{
Hanna Meyer
}

