% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDaytime.R
\name{getDaytime}
\alias{getDaytime}
\title{Classify a scene into day, twilight, night according to its sunzenith}
\usage{
getDaytime(sunzenith)
}
\arguments{
\item{sunzenith}{RasterLayer of sunzenith}
}
\value{
Character with value either "day", "twilight" or "night
}
\description{
Classify a scene into day, twilight, night according to its sunzenith
}
\examples{
sunzenith <-getSunzenith(inpath=system.file("extdata/msg",package="Rainfall"))
getDaytime(sunzenith)
}
\author{
Hanna Meyer
}

