% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSunzenith.R
\name{getSunzenith}
\alias{getSunzenith}
\title{Get sunzenith raster from a msg scene's inputpath}
\usage{
getSunzenith(inpath, type = "rst")
}
\arguments{
\item{inpath}{The path to the msg channels including sunzenith of one scene}

\item{type}{A character string indicating the file type}
}
\value{
The sunzenith raster
}
\description{
Get sunzenith raster from a msg scene's inputpath
}
\examples{
sunzenith <-getSunzenith(inpath=system.file("extdata/msg",package="Rainfall"))
}
\author{
Hanna Meyer
}

