% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ll2cr.R
\name{ll2cr}
\alias{ll2cr}
\title{Get MSG pixel coordinates from geocordiantes}
\usage{
ll2cr(lat, lon, ccoff = 1856, lloff = 1856)
}
\arguments{
\item{ccoff}{Integer coefficient of the scalling function (see page 28, Ref [1])}

\item{lloff}{Integer coefficient of the scalling function  (see page 28, Ref [1])}

\item{latitude}{Numeric}

\item{longitude}{Numeric}
}
\value{
Data frame including column and row of the MSG image
}
\description{
Get MSG pixel coordinates from geocordiantes
}
\examples{
ll2cr(50.7,8.7)
}
\author{
EUMETSAT 2005, 2009. (R implementation: Hanna Meyer)
}
\references{
[1] LRIT/HRIT Global Specification
(CGMS 03, Issue 2.6, 12.08.1999)
 for the parameters used in the program.

 [2] MSG Ground Segment LRIT/HRIT Mission Specific
 Implementation, EUMETSAT Document,
 (EUM/MSG/SPE/057, Issue 6, 21. June 2006).
}

