% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotModelCV.R
\name{plotModelCV}
\alias{plotModelCV}
\title{Plot rfe model or train results based on cross validation}
\usage{
plotModelCV(model, metric = model$metric, tuningValue = "Variables",
  xlim = "minmax", ylim = "minmax")
}
\arguments{
\item{model}{A rfe or train object}

\item{metric}{The metric to visualize. must be stored in the rfe object}

\item{tuningValue}{The tuning value which is depicted on the x axis.
For rfe models default is "Variables", the number of variables.}

\item{xlim}{Numeric vector defining x axis limits. See \code{\link{plot}}.
If xlim == "minmax" these limits are set according to the range of data values.}

\item{ylim}{Numeric vector defining x axis limits. See \code{\link{plot}}.
If ylim == "minmax" these limits are set according to the range of data values.}
}
\description{
function takes an rfe model or a train object and plots the mean metric with
standard deviations to visualize the effect of a tuning parameter (
for rfe model this is the numbers of variables) on the model performance.
}
\note{
returnResamp = "all" must be set in rfe training
}
\examples{
data(rfeModel)
plotModelCV(rfeModel)
}
\author{
Hanna Meyer
}
\seealso{
\code{\link{rfe}}
}

