% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotPredCorr.R
\name{plotPredCorr}
\alias{plotPredCorr}
\title{Plot correlation of predictors}
\usage{
plotPredCorr(predictors, subset = 1:20)
}
\arguments{
\item{predictors}{Either A RasterStack of predictor Variables or a data.frame}

\item{subset}{Indicates the predictors to use}
}
\description{
bases on the corrplot package.
}
\examples{
# stack  msg scenes:
msg_example <-getChannels(inpath=system.file("extdata/msg",package="Rainfall"))
data(rfeModel)
pred<-calculatePredictors(msg_example,model=rfeModel)
plotPredCorr(pred)
}
\author{
Hanna Meyer
}
\references{
Taiyun Wei (2013). corrplot: Visualization of a correlation matrix.
R package version 0.73. http://CRAN.R-project.org/package=corrplot
}

