% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotRfeCV.R
\name{plotRfeCV}
\alias{plotRfeCV}
\title{Plot rfe model results based on cross validation}
\usage{
plotRfeCV(rfeModel, metric = "RMSE")
}
\arguments{
\item{rfeModel}{A rfe object}

\item{metric}{The metric to visualize. must be stored in the rfe object}
}
\description{
function takes an rfe model and plots the mean metric with standard deviations
to visualize the effect of the numbers of variables on the model performance.
}
\note{
returnResamp = "all" must be set in rfe training
}
\examples{
data(BloodBrain)
x <- scale(bbbDescr[,-nearZeroVar(bbbDescr)])
x <- x[, -findCorrelation(cor(x), .8)]
x <- as.data.frame(x)
set.seed(1)
rfeModel <- rfe(x, logBBB,
sizes = c(2, 10,30, 35, 40, 60, 65),
rfeControl = rfeControl(functions = lmFuncs,
method="cv",returnResamp = "all"),method="nnet")
plotRfeCV(rfeModel)
}
\seealso{
\code{\link{rfe}}
}

