% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ppStat.R
\name{ppStat}
\alias{ppStat}
\title{Calculate zonal statistics for cloud patches}
\usage{
ppStat(x, cloudPatches, var)
}
\arguments{
\item{x}{RasterLayer or rasterStack}
}
\description{
Calculate zonal statistics for cloud patches
}

