% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predictRainfall.R
\name{predictRainfall}
\alias{predictRainfall}
\title{Predicts the rainfall rate based on a trained model and MSG data}
\usage{
predictRainfall(model, predictors)
}
\arguments{
\item{model}{The final model from either caret's train or rfe.
Use model$fit$finalModel or model$finalModel to get it}

\item{rainmask}{A raster indicates areas which are not raining with NA values}
}
\value{
A Raster Layer containing predicted rainfall
}
\description{
Predicts the rainfall rate based on a trained model and MSG data
}
\author{
Hanna Meyer
}

