% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/projectMSG.R
\name{projectMSG}
\alias{projectMSG}
\title{Quick&Dirty Method to project MSG data}
\usage{
projectMSG(x, xRaster, yRaster, res = 0.07)
}
\arguments{
\item{x}{A RasterLayer in MSG projection}

\item{xRaster}{A Raster with the same spatial extent as x containing the Latlon
x coordinates as values. Output from the processing routine.}

\item{yRaster}{A Raster with the same spatial extent as x containing the Latlon
y coordinates as values. Output from the processing routine.}

\item{res}{resolution in degree of the projected raster}
}
\value{
x projected to Latlon
}
\description{
Quick&Dirty Method to project MSG data
}
\examples{
#not run
#yRaster=raster(paste(datapath,"/000000000000_00000_ml01danb1_na001_1000_rg01de_003000.rst",sep=""),
 #native = T, crs = "+proj=longlat +datum=WGS84")xRaster=raster(paste(datapath,"/000000000000_00000_ml02danb1_na001_1000_rg01de_003000.rst", sep=""),
 #native = T, crs = "+proj=longlat +datum=WGS84")
 #x=msg_example[[1]]
 #projectMSG(x,xRaster, yRaster)
}
\author{
Hanna Meyer
}

