% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randomScenes.R
\name{randomScenes}
\alias{randomScenes}
\title{Randomly Select a percentage of scenes}
\usage{
randomScenes(sampsize = 0.25, seed = 25, years = c(2010:2010),
  months = c(1:12), days = 1:30, hours = 0:23)
}
\arguments{
\item{sampsize}{proportion of scenes to select}

\item{seed}{A number used as seed}

\item{years}{Start to end year}

\item{days}{start to end days}

\item{hours}{start to end hours}

\item{month}{start to end month}
}
\value{
character vector containing teh dates of the selected scenes
}
\description{
Randomly Select a percentage of scenes
}
\examples{
randomScenes(sampsize=0.01)
}
\author{
Hanna Meyer
}

