% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectralDerivate.R
\name{spectralDerivate}
\alias{spectralDerivate}
\title{Calculation of derivated spectral channels}
\usage{
spectralDerivate(scenerasters, names = c("T0.6_1.6", "T6.2_10.8", "T7.3_12.0",
  "T8.7_10.8", "T10.8_12.0", "T3.9_7.3", "T3.9_10.8"))
}
\arguments{
\item{scenerasters}{A RasterStack of the MSG channels}

\item{names}{Variables to be calculated. Currently
"T0.6_1.6","T6.2_10.8","T7.3_12.0","T8.7_10.8","T10.8_12.0",
"T3.9_7.3","T3.9_10.8" are supported.}
}
\description{
Calculation of derivated spectral channels
}
\examples{
msg_example <-getChannels(inpath=system.file("extdata/msg",package="Rainfall"))
spectralDerivate(msg_example)
}
\author{
Hanna Meyer
}

