% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tempAggregate.R
\name{tempAggregate}
\alias{tempAggregate}
\title{Aggregates MSG images to one hour (mean calculation)}
\usage{
tempAggregate(inpath, outpath = NULL, type = "rst")
}
\arguments{
\item{inpath}{Character string of the path to the MSG scenes.
Only one subfolder and the "/cal" is allowed
(as returned from the msg processing). E.g use this function for one day.
e.g. inpath= /meteosat-images/2010/01/01/ including one folder for every scene
and for every scene the folder "cal"}

\item{outpath}{Character string of the output path}

\item{type}{A character string indicating the file type.
Can be ignored if fromFolder==TRUE}
}
\value{
If no outpath is speciefied, the results are returned to the R
environment as a list of rasterStacks. If outpath is specified,
files are written to disk
}
\description{
Aggregates MSG images to one hour (mean calculation)
}
\author{
Hanna Meyer
}

