% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/textureVariables.R
\name{textureVariables}
\alias{textureVariables}
\title{Calculate selected Texture parameters from clouds based on spectral properties}
\usage{
textureVariables(x, nrasters = 1:nlayers(x), filter = c(3),
  var = c("mean", "variance", "homogeneity", "contrast", "dissimilarity",
  "entropy", "second_moment"), parallel = TRUE, n_grey = 128,
  min_x = NULL, max_x = NULL)
}
\arguments{
\item{x}{A rasterLayer or a rasterStack containing different channels
where clouds are already masked}

\item{nrasters}{A vector of channels to use from x. Default =nlayers(x)}

\item{filter}{A vector of numbers indicating the environment sizes for which
the textures are calculated}

\item{var}{A string vector of parameters to be calculated.
see \code{\link{glcm}}}

\item{parallel}{A logical value indicating whether parameters are calculated
parallely or not}

\item{n_grey}{Number of grey values. see \code{\link{glcm}}}

\item{min_x}{for each channel the minimum value which can occur. If NULL then
the minimum value from the rasterLayer is used.}

\item{max_x}{for each channel the maximum value which can occur. If NULL then
the maximum value from the rasterLayer is used.}
}
\value{
A list of RasterStacks containing the texture parameters for each
combination of channel and filter
}
\description{
Calculate selected Texture parameters from clouds based on spectral properties
}
\details{
This functions fills the glcm function with standard settings used
for the rainfall retrieval
}
\examples{
## example on how to calculate texture from a list of msg channels

#'# stack the msg scenes:
msg_example <-getChannels(inpath=system.file("extdata/msg",package="Rainfall"))

# raster the sunzenith
sunzenith<-getSunzenith(inpath=system.file("extdata/msg",package="Rainfall"))

#calculate texture
result <- textureVariables(msg_example,nrasters=1:3,
var=c("mean", "variance", "homogeneity"))

#plot the results from VIS0.6 channel:
plot(result$size_3$VIS0.6)
}
\author{
Hanna Meyer
}
\seealso{
\code{\link{glcm}}
}

