% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/train4rainfall.R
\name{train4rainfall}
\alias{train4rainfall}
\title{Performs recursive feature selction}
\usage{
train4rainfall(predictors, response, scaleVars = FALSE, sampsize = 0.25,
  threshold = 0.06, nnetSize = 2:5, nnetDecay = c(0.05, 0.07),
  thresholdTune = c(seq(0, 0.1, 0.05), seq(0.12, 0.3, 0.02), seq(0.35, 1,
  0.05)), seed = 20)
}
\arguments{
\item{predictors}{Either a data.frame with each column is one predictor and each
row represents one pixel. Or (if only one scene is used for training)
a RasterStack with one Raster is one Predictor Variable.}

\item{response}{A vector of either Rainfall area or rainfall rates for
the corresponding pixels in predictors. If only one scene is used for model
training, "response" may also be a RasterLayer of the response variable.}

\item{scaleVars}{Center and scale variables?}

\item{threshold}{if response is Rainfall rate: pixels larger than
the threshold are used for rainfall rate training}

\item{nnetSize}{Number of hidden units in nnet}

\item{thresholdTune}{optional threshold tuning. Only if response ="RInfo"}

\item{seed}{Any integer number. Used to produce reproducable results}

\item{nnetDecay.}{Decay value(s) used in nnet training}
}
\value{
A train object. If keepScaling=TRUE a list with the first object is
 the train object and the second object is a data.frame including mean and sd
 values for all predictors which can be used for ensuring same scaling
 with new unknown values.
}
\description{
Performs recursive feature selction
}
\examples{
#' # stack the msg scenes:
msg_example <-getChannels(inpath=system.file("extdata/msg",package="Rainfall"))

# raster the sunzenith
sunzenith<-getSunzenith(inpath=system.file("extdata/msg",package="Rainfall"))

#get Date
date <- getDate(system.file("extdata/msg",package="Rainfall"))

response <- raster(system.file("extdata/radar",
"201007121650_radolan_SGrid.rst",package="Rainfall"))

#get optimal variables from rfe model
data(rfeModel)
pred<-calculatePredictors(msg_example,model=rfeModel,date=date)

train4rainfall(pred,response,sampsize=0.1)
}
\author{
Hanna Meyer
}
\references{
train Function in the caret package
}
\seealso{
\code{\link{rfe}}
}

