% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validation of model results}
\usage{
validate(obs, pred, type = "regression", ...)
}
\arguments{
\item{obs}{A RasterLayer of observed rainfall areas or rainfall rates}

\item{pred}{A RasterLayer of predicted rainfall areas or rainfall rates}

\item{type}{A character string indicating the type of the model output.
Either "classification" or "regressione"}
}
\value{
A data.frame containing the validation statistics
}
\description{
The calculations base on the regressionStats or
classificationStats of the Rsenal package
}
\author{
Hanna Meyer
}
\seealso{
\code{\link{classificationStats}}, \code{\link{regressionStats}}
}

