% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/varFromRfe.R
\name{varFromRfe}
\alias{varFromRfe}
\title{Extract variables from rfe model}
\usage{
varFromRfe(rfeModel, useOptimal = TRUE)
}
\arguments{
\item{rfeModel}{An object of class rfe}

\item{useOptimal}{Logical. Use the optimal variables from rfe or those
less variables which lead to a model performance within one sd of the
optimal model?}
}
\value{
A list of variables from the rfe object which can be directly used
in calculatePredictors
}
\description{
This is an internal function used by calculatePredictors
}
\examples{
data(rfeModel)
varFromRfe(rfeModel)
}
\author{
Hanna Meyer
}

