% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/writeToFile.R
\name{writeToFile}
\alias{writeToFile}
\title{Writes a stack of MSG channels to a new file with respect to file
name conventions}
\usage{
writeToFile(scenerasters, date, outpath)
}
\arguments{
\item{scenerasters}{Raster stack of MSG scenes originally retrieved by
\code{\link{getCahnnels}}}

\item{date}{Date of the scene. (see \code{\link{getDate}})}

\item{outpath}{Path to where files are to be written}
}
\value{
Nothing is returned to the R environment. New files are stored on disk.
}
\description{
Writes a stack of MSG channels to a new file with respect to file
name conventions
}
\details{
maimly used as internal function of \code{\link{tempAggregate}}
}
\author{
Hanna Meyer
}

