% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculatePredictors.R
\name{calculatePredictors}
\alias{calculatePredictors}
\title{Calculate predictors from MSG cloud masked data}
\usage{
calculatePredictors(scenerasters, sunzenith, spectral, texture, pptext, zonstat,
  shape, further = c("sunzenith", "jday"), date)
}
\arguments{
\item{scenerasters}{A raster stack of cloud masked MSG scenes with non
clouded areas were set to NA}

\item{sunzenith}{A raster of the sun zenith values}

\item{spectral}{A character vector indicating the msg channels to be included.
Possible values: "VIS0.6","VIS0.8","NIR1.6","IR3.9","WV6.2","WV7.3","IR8.7",
"IR9.7","IR10.8","IR12.0","IR13.4","T0.6_1.6","T6.2_10.8","T7.3_12.0",
"T8.7_10.8","T10.8_12.0", "T3.9_7.3","T3.9_10.8"}

\item{texture}{data frame of all spectral and texture combinations which are
to be calculated. (Tip: Use expand.grid to create this data.frame)}

\item{pptext}{data frame of all spectral and texture combinations which are
to be calculated for teh overall cloud entity.
 (Tip: Use expand.grid to create this data.frame)}

\item{zonstat}{data frame of all spectral and zonal stat combinations
(min,max,mean or sd) which are to be calculated for the overall cloud entity.
(Tip: Use expand.grid to create this data.frame)}

\item{shape}{geoemtry variables which should be included. Possible values
are "Ar",SI","CA","Ur","CAI","PAR","distEdges","Re","Ru","OIC",
CI1","CO1","CI2","CO2","CCI1","CCI2","CO","SHD","C1","E",
"TR","CR","C2","FR","EI","SF1","GSI","SF2","C3","SF3"}

\item{further}{a character vector including Currently "jday" and/or
"sunzenith" which will also be used as variables.
see \code{\link{geometryVariables}} and \code{\link{borgIndices}}
for description of the variables.}

\item{date}{Date of the msg scene in format yyyymmddhhmm. Only imprtant if
the day of the year (jday) is calculated (see param "further").}
}
\description{
Calculate predictors from MSG cloud masked data
}
\examples{
#list msg raster in the example folder:
scenes<-list.files(system.file("msg",package="Rainfall"),
pattern=".rst$")

# name the variables (according file name convention)
#which are to be used:
x=c("ca02p0001","ca02p0002","ca02p0003","ct01dk004","ct01dk005",
"ct01dk006","ct01dk007","ct01dk008","ct01dk009","ct01dk010","ct01dk011")

# raster the sunzenith raster which is defined as "ma11" according to
#file name conventions:
sunzenith<-raster(system.file("msg",
scenes[substr(scenes,20,23) =="ma11"],package="Rainfall"))

# stack the msg scenes:
msg_example <-  stack(system.file("msg",
scenes[substr(scenes,20,28)\%in\%x],package="Rainfall"))

date <- substr(scenes[1],1,12)

# set non clouded areas to NA:
msg_example=reclassify(msg_example, cbind(-99,NA))

# name the msg channels:
names(msg_example)<-c("VIS0.6","VIS0.8","NIR1.6","IR3.9",
"WV6.2","WV7.3","IR8.7","IR9.7","IR10.8","IR12.0","IR13.4")

#calculate variables (takes some time...)
pred <- calculatePredictors(msg_example,
sunzenith,
spectral=c("VIS0.6","VIS0.8","NIR1.6","T0.6_1.6","T6.2_10.8"),
texture=expand.grid(c("VIS0.8","NIR1.6","T6.2_10.8"),
c("variance", "contrast")),
pptext=expand.grid("T3.9_10.8",c("variance", "contrast","mean")),
shape=c("Ar","CAI","SI","CI1"),
zonstat=data.frame("spec"=c("VIS0.8","VIS0.8","WV7.3","T6.2_10.8"),
var=c("mins","sds","mins","maxs")),
date=date)
print(pred)
}

