% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubset.R
\name{createSubset}
\alias{createSubset}
\title{Create a training subset from data}
\usage{
createSubset(predictors, response, out = "Rain", threshold = 0.06,
  sampsize = 0.01, seed = 20)
}
\arguments{
\item{predictors}{Either a data.frame with each column is one predictor and each
row represents one pixel. Or (if only one scene is used for training) 
a RasterStack with one Raster is one Predictor Variable.}

\item{response}{A vector of either Rainfall area or rainfall rates for
the corresponding pixels in predictors. If only one scene is used for model 
training, "response" may also be a RasterLayer of the response variable.}

\item{out}{Either Rain or RInfo indicating weather rainfall rates
or rainfall areas should be used.}

\item{threshold}{if response is Rainfall rate: pixels larger than
the threshold are used for rainfall rate training}

\item{sampsize}{Numeric value between 0 and 1 indicating the percentage of
data points to be returned}

\item{seed}{see \code{\link{set.seed}}}
}
\value{
A list with two entries: predictors and response
}
\description{
Bases on createDataPartition from the caret package.
}
\author{
Hanna Meyer
}
\seealso{
\code{\link{createDataPartition}}
}

