% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filterStats.R
\name{filterStats}
\alias{filterStats}
\title{Calculate statistics of spatial filters of MSG bands}
\usage{
filterStats(scenerasters, size = 3, var = c("min", "mean", "max", "sd"),
  na.rm = TRUE)
}
\arguments{
\item{scenerasters}{A raster stack of cloud masked MSG scenes with non
clouded areas were set to NA}

\item{size}{Integer of the window size}

\item{var}{Variables to be calculated. Supported are min, mean, max, sd}

\item{na.rm}{Should NA values from scenerasters be masked from the result?}
}
\description{
Calculate statistics of spatial filters of MSG bands
}
\author{
Hanna Meyer
}

