% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geometryVariables.R
\name{geometryVariables}
\alias{geometryVariables}
\title{Calculate selected geometry variables for clouds}
\usage{
geometryVariables(x)
}
\arguments{
\item{x}{A rasterLayer containing NA for non clouds and any value for
clouded areas}
}
\value{
A list of RasterStacks containing the texture parameters for each
combination of channel and filter
}
\description{
Calculate selected geometry variables for clouds
}
\examples{
msg_example <-  raster(system.file("msg",
"201010081250_mt09s_ct01dk009_m1hct_1000_rg01de_003000.rst",
package="Rainfall"))

# set non clouded areas to NA:
msg_example=reclassify(msg_example, cbind(-99,NA))

#calculate geometry Variables
geometry <- geometryVariables(msg_example)
plot(geometry[[1:5]])
}
\author{
Hanna Meyer
}
\seealso{
\code{\link{SDMTools}}, \code{\link{clump}}, \code{\link{borgIndices}}
}

