% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glcmPerPatch.R
\name{glcmPerPatch}
\alias{glcmPerPatch}
\title{Calculate selected Texture parameters for overall cloud entities
 based on spectral properties
 @param x 
 @return a raster stack of the texture parameters for each entity
 @author Hanna Meyer
 @export glcmPerPatch
 @seealso \code{\link{calculateTexture}}, \code{\link{glcm}}
 @examples
 # stack the msg scenes:
msg_example <-getChannels(inpath=system.file("extdata/msg",package="Rainfall"))}
\usage{
glcmPerPatch(x, patches, nrasters = 1:nlayers(x), var = c("mean",
  "variance", "homogeneity", "contrast", "dissimilarity", "entropy",
  "second_moment"), n_grey = 32)
}
\description{
#calculate cloud patches
geometry <- geometryVariables(msg_example[[1]],var="cloudPatches")
}
\details{
#calculate glcm per patch
glcmPerPatch(msg_example[[1]],geometry$cloudPatches,var=c("mean"))
}

