% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/glcmPerPatch.R
\name{glcmPerPatch}
\alias{glcmPerPatch}
\title{Calculate selected Texture parameters for overall cloud entities
 based on spectral properties}
\usage{
glcmPerPatch(x, patches, nrasters = 1:nlayers(x), var = c("mean",
  "variance", "homogeneity", "contrast", "dissimilarity", "entropy",
  "second_moment"), n_grey = 32)
}
\arguments{
\item{x}{}
}
\value{
a raster stack of the texture parameters for each entity
}
\description{
Calculate selected Texture parameters for overall cloud entities
 based on spectral properties
}
\examples{
# stack the msg scenes:
msg_example <-getChannels(inpath=system.file("extdata/msg",package="Rainfall"))

#calculate cloud patches
geometry <- geometryVariables(msg_example[[1]],var="cloudPatches")

#calculate glcm per patch
glcmPerPatch(msg_example[[1]],geometry$cloudPatches,var=c("mean"))
}
\author{
Hanna Meyer
}
\seealso{
\code{\link{calculateTexture}}, \code{\link{glcm}}
}

