% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ll2cr.R
\name{ll2cr}
\alias{ll2cr}
\title{Get MSG pixel coordinates (column/rows) from geocordiantes}
\usage{
ll2cr(latlon, ccoff = 1856, lloff = 1856)
}
\arguments{
\item{latlon}{data.frame with first column =latitude and second column is
longitude specified as numeric.}

\item{ccoff}{Integer coefficient of the scalling function (see page 28, Ref [1])}

\item{lloff}{Integer coefficient of the scalling function  (see page 28, Ref [1])}
}
\value{
Data frame including column and row of the MSG image
}
\description{
Get MSG pixel coordinates (column/rows) from geocordiantes
}
\examples{
#for a single coordinate
ll2cr(data.frame(50.7,8.7))

#for more than one coordinate
latlon=data.frame("lat"=c(50.7, 53.8),"lon"=c(8.7, 11.2))
ll2cr(latlon)
}
\author{
Hanna Meyer based on the FORTRAN Routine of EUMETSAT 2005, 2009.
}
\references{
[1] LRIT/HRIT Global Specification
(CGMS 03, Issue 2.6, 12.08.1999)
 for the parameters used in the program.

 [2] MSG Ground Segment LRIT/HRIT Mission Specific
 Implementation, EUMETSAT Document,
 (EUM/MSG/SPE/057, Issue 6, 21. June 2006).
}

