% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgstatmatch.R
\name{msgstatmatch}
\alias{msgstatmatch}
\title{Match temporally aggregated (1h) MSG data with station data}
\usage{
msgstatmatch(stationpath, MSG_extract, UTC = 0)
}
\arguments{
\item{stationpath}{path to the csv files of the 1h weather station data}

\item{MSG_extract}{result from extractMSG. The first column is the date, the 
second column is the station name. The MSG data of the corresponding time and
location are organized in the subsequent columns.}

\item{UTC}{numeric. Time Zone of the station data specified in UTC hours from the UTC}
}
\description{
Match temporally aggregated (1h) MSG data with station data
}
\author{
Hanna Meyer
}

