% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ppStat.R
\name{ppStat}
\alias{ppStat}
\title{Calculate zonal statistics for cloud patches}
\usage{
ppStat(scenerasters, cloudPatches, var = c("mean", "sd", "min", "max"))
}
\arguments{
\item{scenerasters}{RasterLayer or rasterStack of spectral channels}

\item{cloudPatches}{A Raster of cloud entities calculated
by \code{\link{geometryVariables}}}

\item{var}{vector containing values of "mean","sd","min","max"}
}
\description{
Calculate zonal statistics for cloud patches
}
\examples{
msg_example <-getChannels(inpath=system.file("msg",package="Rainfall"),
channels="VIS0.8")

#calculate geometry Variables
geometry <- geometryVariables(msg_example,var=c("cloudPatches"))

ppStat(msg_example,geometry$cloudPatches,var=c("mean","sd"))
}

