% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predictRainfall.R
\name{predictRainfall}
\alias{predictRainfall}
\title{Predicts the rainfall rate based on a trained model and MSG data}
\usage{
predictRainfall(model, inpath = NULL, sceneraster = NULL, rainmask = NULL,
  sunzenith = NULL, date = NULL, useOptimal = TRUE,
  scaleparam = model$scaleParam, min_x = NULL, max_x = NULL,
  type = "tif")
}
\arguments{
\item{model}{The final model from either caret's train or rfe.
Use model$fit$finalModel or model$finalModel to get it}

\item{inpath}{Path to the MSG data}

\item{sceneraster}{If no inpath is specified:
The Meteosat data from which rainfall should be predicted.
Load them with getChannels}

\item{rainmask}{A raster indicates areas which are not raining with NA values}

\item{useOptimal}{if model is a rfe object: Logical. Use the optimal variables
from rfe or those less variables which lead to a model performance within one
sd of the optimal model?}

\item{scaleparam}{A data.frame created with \code{\link{calcScalingStats}}}

\item{type}{If inpath!=NULL then the data type of the MSG raster}

\item{If}{no inpath is specified: sunzenith optional.
Only needed if included in the predictor variables}

\item{If}{no inpath is specified: date optional.
only needed if jday is included in the predictor variables}
}
\value{
A Raster Layer containing predicted rainfall
}
\description{
Functions calculates predictors which are required by the model
and uses them for prediction
}
\examples{
# stack the msg scenes:
msg_example <-getChannels(inpath=system.file("extdata/msg",package="Rainfall"))

reference <- raster(system.file("extdata/radar",
"201007121650_radolan_SGrid.rst",package="Rainfall"))
values(reference)[values(reference<0.06)]=NA

data(rfemodel)

#predict on the new scene (don't expect good results from the small model!)
pred<-predictRainfall(model=rfeModel, sceneraster=msg_example, rainmask=reference)
validate(obs=reference,pred=pred)
}
\author{
Hanna Meyer
}

