\docType{package}
\name{Rainfall-package}
\alias{Rainfall-package}
\alias{rainfallpackage}
\title{Functions for training and prediction of an MSG based rainfall retrieval}
\description{
Rainfall
}
\author{
Hanna Meyer\cr
\emph{Maintainer:} Environmental Informatics \email{admin@environmentalinformatics-marburg.de}
}
\keyword{package}

