\name{rainfall-package}
\alias{rainfall-package}
\alias{rainfall}
\docType{package}
\title{
\packageTitle{rainfall}
}
\description{
\packageDescription{rainfall}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rainfall}
\packageIndices{rainfall}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{rainfall}

Maintainer: \packageMaintainer{rainfall}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
