% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rfe4rainfall.R
\name{rfe4rainfall}
\alias{rfe4rainfall}
\title{Performs recursive feature selction}
\usage{
rfe4rainfall(predictors, response, sampsize = 0.25, threshold = 0.06,
  varSize = c(1:5, seq(10, ncol(predictors), 10)), nnetSize = seq(2,
  ncol(predictors), 8), nnetDecay = 0.05, seed = 20)
}
\arguments{
\item{predictors}{Either a data.frame with each column is one predictor and each
row represents one pixel. Or (if only one scene is used for training)
a RasterStack with one Raster is one Predictor Variable.}

\item{response}{A vector of either Rainfall area or rainfall rates for
the corresponding pixels in predictors. If only one scene is used for model
training, "response" may also be a RasterLayer of the response variable.}

\item{threshold}{if response is Rainfall rate: pixels larger than
the threshold are used for rainfall rate training}

\item{varSize}{integer vector indicating the numbers of
variables to consider in rfe.}

\item{nnetSize}{Number of hidden units in nnet}

\item{seed}{Any integer number. Used to produce reproducable results}

\item{nnetDecay.}{Decay value(s) used in nnet training}
}
\value{
A rfe model
}
\description{
Performs recursive feature selction
}
\details{
Predictors are centered and scaled according to mean and sd values.
If the day of the year is used as predictor, this variable is
scaled considering max=365 and min=1
}
\examples{
# stack the msg scenes:
msg_example <-getChannels(inpath=system.file("extdata/msg",package="Rainfall"))

# raster the sunzenith
sunzenith<-getSunzenith(inpath=system.file("extdata/msg",package="Rainfall"))

#get Date
date <- getDate(inpath)

#calculate variables (takes some time...)
pred <- calculatePredictors(msg_example,
sunzenith,
spectral=c("VIS0.6","NIR1.6","T0.6_1.6"),
texture=expand.grid(c("NIR1.6","T6.2_10.8"),
c("variance", "contrast")),
 pptext=expand.grid("T3.9_10.8",c("variance","mean")),
 shape=c("Ar","CAI","SI","CI1"),
 zonstat=data.frame("spec"=c("VIS0.8","VIS0.8","T6.2_10.8"),
 var=c("min","sd","max")),
 further=NULL,
 date=date)

response <- raster(system.file("extdata/radar",
"201010081250_radolan_SGrid.rst",package="Rainfall"))

#Train small rfe model with 0.1\% of the pixels (takes around 1 minute...)
rfeModel <- rfe4rainfall(predictors,
response,
sampsize=0.001)

# Show results:
print(rfeModel)
plotRfeCV(rfeModel)
predictors(rfeModel)
plot(varImp(rfeModel$fit,scale=TRUE))
}
\author{
Hanna Meyer
}
\references{
rfe Function in the caret package
}
\seealso{
\code{\link{rfe}}
}

