% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scaleByValue.R
\name{scaleByValue}
\alias{scaleByValue}
\title{Scales Predictors using mean and sd from lookup table}
\usage{
scaleByValue(x, scaleTable)
}
\arguments{
\item{x}{RasterStack of the scene which should be scaled or data.frame}
}
\value{
scaleTable Data.frame created with \code{\link{calcScalingStats}}
}
\description{
Scales Predictors using mean and sd from lookup table
}
\author{
Hanna Meyer
}

