% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/varFromRfe.R
\name{varFromRfe}
\alias{varFromRfe}
\title{Extract variables from rfe model}
\usage{
varFromRfe(rfeModel)
}
\arguments{
\item{rfeModel}{An object of class rfe}
}
\value{
A list of variables from the rfe object which can be directly used
in calculatePredictors
}
\description{
This is an internal function used by calculatePredictors
}
\examples{
data(rfeModel)
varFromRfe(rfeModel)
}
\author{
Hanna Meyer
}

